requirements.txt
Django==2.1.7
djangorestframework==3.9.2
pip==10.0.1
psycopg2==2.7.7
pytz==2018.9
PyYAML==3.13
setuptools==39.1.0